requirements.txt
numpy>=1.18.0
matplotlib>=3.1.0
pygame>=2.0.0
PyOpenGL>=3.1.0
PyOpenGL_accelerate>=3.1.0